requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
cryptography==41.0.7
mediapipe==0.10.7
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
